requirements.txt
django==4.1.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
django-nested-admin==4.0.2
psycopg2-binary==2.9.5
Pillow==9.3.0
python-dateutil==2.8.2
drf_yasg
